fix compile warning
git-svn-id: http://192.168.20.11/sequoiadb/trunk@27716 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dms/dmsStorageIndex.cpp
@@ -230,8 +230,8 @@ namespace engine
             }
 
             /// update last lsn
-            if ( ~0 == lastLSN ||
-                 ( ~0 != tmpLSN && lastLSN < tmpLSN ) )
+            if ( (UINT64)~0 == lastLSN ||
+                 ( (UINT64)~0 != tmpLSN && lastLSN < tmpLSN ) )
             {
                lastLSN = tmpLSN ;
             }
